+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2.11.17
+
+#### Dependency upgrades
+* Bump spring-security-oauth2 from 2.5.1.RELEASE to 2.5.2.RELEASE
+* Bump spring-boot-starter version from 2.6.6 to 2.6.7
+
+
 ## 2.11.16
 - [java-security] [spring-security] JwtSignatureValidator improvements:
   - Only identity service requires `x-zone_uuid` header for token keys retrieval
@@ -12,7 +19,7 @@ Introduces with `spring-security-compatibility` a compatibility module that prov
 #### Dependency upgrades
 * Bump spring-boot-starter-parent version from 2.5.2 to 2.6.6 (only in samples)
 * Bump reactor-core from 3.4.16 to 3.4.17
-
+* Bump spring.security.version from 5.6.2 to 5.6.3
 
 ## 2.11.15
 [spring-xsuaa][spring-security]